LL: build_and_checks.sh with argument for applying to other folders, re? build_readme.sh hardening when no README_printable.md to avoid emptying README.md.

+++ build_readme.sh
@@ -32,10 +32,15 @@ then
   cd "$1"
 fi
 
-sed -Ez "$sed_expression" README_printable.md > README_temp1.md
-sed -Ez "$sed_expression" README_temp1.md > README_temp2.md
-sed -Ez "$sed_expression" README_temp2.md > README_temp3.md
-sed -Ez "$sed_expression" README_temp3.md > README.md
-rm README_temp1.md README_temp2.md README_temp3.md
+if [[ -f README_printable.md ]];
+then
+  sed -Ez "$sed_expression" README_printable.md > README_temp1.md
+  sed -Ez "$sed_expression" README_temp1.md > README_temp2.md
+  sed -Ez "$sed_expression" README_temp2.md > README_temp3.md
+  sed -Ez "$sed_expression" README_temp3.md > README.md
+  rm README_temp1.md README_temp2.md README_temp3.md
+else
+  echo "No file README_printable.md"
+fi
 
 popd
